requirements.txt
Flask
gunicorn
sqlalchemy
google-cloud-aiplatform
psycopg2-binary  # or mysql-connector-python for MySQL